Adjusts right click behavior in factory build menu (#5861)
Right click on the unit in build menu, in quota mode, will decrease the queue if quota was at 0.
Right click on the unit in build menu, in queue mode, will decrease the quota if queue was at 0.

It will allow to cancel queues or quotas without switching to other mode.