Napalm being tricked by very-high LLT midpoint (#5201)
The basic napalm units don't get to stack their damage (as much) anymore, so missing LLTs that it's their role to destroy hurts their effectiveness. Players noticed pretty fast, even though this behavior is not new.

A unit's midpoint is what makes it visible, so turrets have elevated midpoints. This reveals them to any unit they can shoot (and with enough sight radius). The LLT's midpoint is located along the axis of its laser weapon at the very top of the unit, so it has to be engulfed completely to take napalm damage.

Add a check for the unit's base point, in addition to the midpoint. The new test is more involved, since units can be in any orientation, e.g. when climbing or sent flying.

For optimization, a lidded cylindrical test is done before the spherical test. That gave me an opportunity to lift up the base slightly when on water, to protect subs from napalm damage. (If underwater units need immunity to the effect then we can enforce that, sure, but that's not what this PR was for.)

Then, if the midpoint is outside of the damage area, the damage CEG is placed at the intersection of the ray from mid->base point and the sphere of the AoE. This elevates the CEG without letting it spawn too far away from the source of the effect.