fix maxvalue -> maxscale, limits max slowdown from water
Set to 1.4 as roughly the inverse of the previous maxValue. A maxScale less than 1 does not make sense -- effectively always speeding up the unit -- so I have truly no idea what was going on, here.

That seems to cap the depth that slows the unit to 1.4 / 0.03 = 47 elmos, which is a moderate depth. This has balance implications for the vehicles and bots using the generic depth mod params and with a max depth >= 47.

Very large units probably should have a higher depth for their maxScale cutoff than this.