Grid build for pregame (#6041)
adds grid, square, line, and surround build key combos for pregame
build. Works with quickstart custom square colors. See attached video.

Testing steps:
- play game normally or  with quickstart
- if using quickstart, pay attention to purple square outlines around
buildings when they're in the budget.
- otherwise, try CTRL + Shift for surround build
- hold shift only and click+drag to do line builds
- hold shift and click and it'll append another build to the queue
- CTRL + Shift + Alt for square build (empty space in middle)
- Shift + Alt for grid build
- while Shift + Alt, press Z or X to increase or decrease spacing

abuse it, try to break it.
## known suckiness
the performance of this widget is ass. Prior attempts with more major
overhauls barely yielded performance improvements. The only real option
for improvement is to use ghost VBO's, square vbo's, range vbo's etc
instead of drawing gl.shapes. This is very complicated and I don't wanna
do it. Shaders are pain.
as it is, it performs just fine so long as you don't try to cover an
absurd area (like an entire 12x12 map) with a grid build.

This was made in tandem with an AI assistant. I've read and tested the
code myself. See video (which I forgot to do line build on. Whoops.
Trust me bro, it totally works.)
https://youtu.be/YwpeycDdtz8