fix: pregame substitution
getEquivalentUnitDefID was a lie, it was refactored to deal in names only, which made a lot of sense for blueprints but did break the pregame substitution, which dealt strictly in unitDefID. I had simplified this six months ago, forgot about blueprint substitution being mixed up in the pregame logic, then failed to test that aspect of substitutions. Whoops.