check for ultrawide units much earlier
This avoids waiting on the buggeroff radius to increase to detect units with a very large radius. This was an issue with my test of the Armada T2 bot lab, which has a larger footprint than its radius (todo: should detect blocking via footprint, not unit radius), and Fatboys, which are both large and tremendously slow. The earlier we get the message to them, the better.

Also, I shortened the var name interferingUnitID => unitID but maybe that's an issue of personal taste? Happy to change it back, idc.