pregame gridmenu text outline fix (#6085)
There's a weird issue in the pregame where in gridmenu and only in
gridmenu the outline around text appears and disappears at random. Seems
related to whenever certain updates happen and I'm not exactly sure why.
It's totally fine during the game.

Explicitely adding the outline when creating the text seems to do the
trick.

Before: 

https://github.com/user-attachments/assets/5385fbd3-ed5d-4140-af53-02fb7ccb3a2b

After

https://github.com/user-attachments/assets/8df598cc-6c99-498b-b907-09176f8da7e1