New Gadget `unit_waterspeedmultiplier` To Selectively Adjust Land and Sea Traversal Speeds Seperately (#5814)
### Work done
Introduces a new gadget called `unit_waterspeedmultiplier` which takes
the base maxSpeed value assigned to a unit inside their unitDef and uses
a custom parameter called `speedfactorwater` (default set to 1.0) to
determine what percent the unit should speed up or slow down over water.
It is only useful for units that can traverse across both land and sea,
such as hovercrafts and Platypuses, which help slow them down when they
are on land such that other factors for their balance can be made
stronger in compensation for reduced traversal speed on land compared to
water (where ships require better stats as a base as standard to fight
against effectively).

Update, 9/22/2025:
Now performs factor to turn rate, acceleration, and deceleration values
of the unit as well. Turn rate is affected at half the effectiveness of
the given factor, accel/decel at 3/4 the effectiveness. A unit with 10
speed, 5 turn rate, and 1 accel/decel, under a `speedfactorwater` of
0.5, has 5 speed, 3.75 (5-1.25=3.75) turn rate, and 0.75 accel/decel.

This gadget does not affect how units that traverse *underwater* behave.
Only over water.
This gadget does nothing for air units even with a customparam applied. 

#### Addresses Issue(s)
- Previously, there was no way to force units to move at different
speeds on land and across water, not even with moveDefs. Introduces ways
to help better balance units that cross both land and water without
being locked onto one specific traversal speed.

#### Setup
Provide any unit with the customParam field `speedfactorwater` and set
it to any value. Right now, only units with HOVER movedefs are eligible.

#### Test steps
- [ ] Add `speedfactorwater` value from -10.0 to 10.0 (there is no
actual limit)
- [ ] Check and see if the unit is faster or slower on water
appropriately
- [ ] Check and see if the unit properly transitions to the
faster/slower assigned speed when crossing the threshold between land
and water.
- [ ] Check and see if the unit properly transitions when being given
various commands and various combinations and frequency of them to see
if the speed transition is ever interrupted.

#### Video:
Below, you can see the Triton being supplied with a `speedfactorwater`
of 2, meaning that it travels at 60 speed on water and 30 speed on land.


https://github.com/user-attachments/assets/01ab2c74-23c4-4840-9a11-60b1c3d1865e

---------

Co-authored-by: efrec <103912894+efrec@users.noreply.github.com>
Co-authored-by: efrec <ecfrechette@gmail.com>