(feat): Added unsynced widget API for Lua analytics via Spring.SendLuaUIMsg (#6012)
Also adds new "widget report" that uses this new API to report custom widget usage

This PR is fairly optimized as it spreads the workload over multiple frames as much as possible.

The limit for a message has been said to be 32k. With my 26 custom widgets the message was below 6kb. But just in case the custom widgets array is truncated to 50 widgets (and reported as truncated).