fix: guidance laser paints targets only when firing (#6208)
### Work done

Adds a little util function to unit_custom_weapons_behaviors, refactors
the `guidance` speceffect a bit, and then modified it to provide target
guidance only when the targeting laser is firing.

This prevents an odd cheese with the Legion Medusa allowing it to target
with precision beyond its ordinary range by switching targets in
mid-flight. I'm sure there is some way to exceed the range, still, but
it would be wildly inaccurate, most likely.

Tested by switching between targets in all kinds of ways with a Medusa.

---------

Co-authored-by: efrec <quantumscrewdriver@gmail.com>