Luaify and guiafy reactive armor (#6202)
### Work done

Replace BOS/COB reactive armor with a Lua gadget and update all unit
scripts to match the new gadget. The gadget handles damage to the armor
health and the armor's recovery, including an out-of-combat timer.

The gadget has some requirements that might not hold out forever but are
fine for now. There have to be an equal number of `ReactiveArmorBreakX`
and `ReactiveArmorRestoreX` methods on the unit, and the unit script is
responsible for setting armored/unarmored states. I'm not overly
attached to either of these details.

There are subtle differences between this implementation and the current
one. This one freezes, rather than resets, the recovery progress when
the unit reenters combat (by taking damage). I think that's sensible,
but maybe it would cause players to be surprised when a unit recovers
faster than it "should". This is also slightly, but not entirely,
`/luarules reload` safe since units do not recall their out-of-combat
timers. These are the only functional differences.

The visuals are slightly different, also. Armor pieces are gradually
lost as reactive armor's health is worn down. They are still restored
gradually, piece by piece. When a unit takes a lot of damage at once,
multiple pieces can be lost at once, so I reduced the chance that each
armor piece forms a projectile (with a kind of hack way of randomizing
that).

Finally, there are GUI changes. The reactive armor units got health bars
(which are just smuggled shield bars, at the moment) when armored but
not at full armor hp. Reactive armor stats are also displayed in the
tooltip window on the unit (select unit -> i -> check Open/Closed health
stats).

#### Addresses Issue(s)

- indiscernible game mechanics and game state

#### Test steps

- Spawn a few each of: Karkinos (legkark), Telchine (legamph), Phalanx
(legshot)
- Bust them up with some enemy Pawns or something
- Watch their armor go down, break, and recover
- Watch a unit with non-broken armor recover its armor, also

---------

Co-authored-by: efrec <quantumscrewdriver@gmail.com>