Legion Hotfix 11.25.7 - Naval Unit Fixes/Adjustments (#6272)
### Work Done
* Octeres' second turret, when trying to aim between -125 deg and 165
degrees, will only be able to shoot when the pitch is 30 or above.
Otherwise, it does not aim at all. It will also try to turn to the
middle point first if the angle difference is such that it will need to
turn into the hull, to avoid clipping into the hull itself.
* The reloadtimer has been fixed for the Octeres.
* Shipyard buildpoint moved forward a bit, accommodates all ships
without clipping now.
* Iapetus midpoint fixed, was too far back before.
* Syracusia's turret placement was raised a bit so it can reliably hit
enemy units a bit more.
* Argonaut's torpedoes will no longer collide with friendly units,
removing friendly killing issue.

---------

Co-authored-by: Zecrus2 <72325649+Zecrus2@users.noreply.github.com>