add non-repulsor shield support to cluster munitions
We are still detecting shield hits only on direct contact, so this code, effectively, ignores shields that are a tiny distance away from a direct hit. This still causes some cluster shots to sneak through. I think this is a "sufficient" solution and gives good performance but I'll likely sweep through in a couple months once it bothers me enough and revamp hit detection.