feat: configurable swerve when guidance speceffect loses targeting (#6302)
### Work done

This adds `guidance_miss_radius` to the Legion Medusa to restore,
partially, its previous behavior when it lost a target. Previously,
missiles would go haywire without guidance, though the targeted location
was repeatable enough to be mildly abusable (with a lot of cleverness,
very much so).

This new "swerve" behavior has a configurable radius (which can be zero)
and picks a random direction per-missile.

When the target unit or the Medusa dies, the remaining missiles spread
out and cover a small area around the last targeted position, which
looks a little better than all missiles striking the exact same
position.

This looks especially nice when destroying one target mid-volley and
switching to another. The missile burst spreads out slightly, then snaps
onto the new direction, which just looks very cool.

Since missiles immediately go off-trajectory a small amount, when an
allied unit blocks the targeting laser, the Medusa's missiles can
sometimes swerve around them and avoid the friendly. This is pretty
nice, actually.

When terrain is blocking, instead, sometimes (rarely) a missile will
point itself up and manage to clear it, if it was a tiny bump. I think
this is negligible.

---------

Co-authored-by: efrec <quantumscrewdriver@gmail.com>