let gadgets try and fail to remove the selvol, actually
I had this slightly backwards so got it confused in the last commit. We want gadgets to be able to restore the volume so they can do searches via ray trace. But we don't want them to remove it because then we would have to check for internal consistency in the update loop. Just make things consistent here, at the site of the call.