Quickstart refund rework (#6341)
this PR makes it so that if you:
- have placed at least 1 build
- have less than 800 budget remaining
then you can have the remaining budget refunded as metal.

Also makes the quickstart base-generation range dependent on budget
size. Range reduced to 435 in allignment with zecrus' observation that
many maps mex spacing are based on the range of an LLT. Not worth adding
a comment, because maps aren't necessarily coupled to LLT range.