breaking armor resets its restore time to max
And armor damage/repair each shift the restore time in proportion, as well. Anything, evidently, to make the doArmorDamage function even longer and harder to read.

This patches out the "surprise" element I mentioned in the initial PR but didn't make much noise about. After this change, even without knowing the exact mechanics but knowing "15 second restore time" (for legkark), I didn't encounter any moments when armor was lost or regained and that was outside of my expectation at all.