Fix scouts not being considered combat units (#6366)
For area-select, turns out that armpeep and corfink are considered radar
units due to their characteristics. This adds the customparam that
forces them to be selectable as though they were combat units as before
the recent update that made jammers/radar units toggleable in area
selection.

Also, discovered the customparam didn't work for all kinds of units due
to the order in which the conditions were checked. Fixed that too so now
the customparam is indeed authoratative.

Went through and tested all the filtering builders, buildings, jammers,
radars, antinuke, and now including the t1 air scouts and all seems to
work correctly now. Didn't test every unit in the game.