Add Missing API for blocking buildoptions (#6362)
I could have sworn I made this already - guess it was a lost branch.
Anyway, this adds an API used by tech blocking experimental feature to
block building T2 and T3 labs. It does this by adding a list of reason
to block a particular UnitDefID. Remove the reason, unblock it.

Tested both legacy and gridmenu layouts in a variety of blocking
situations, she works.