Show/Choose ally AI spawns (#6342)
- allows you to see where the AI will spawn
- any allied player can choose where to spawn an ally AI
- right click a ai cone to de-place them

~~This was made with AI assistance, I'm still combing through it
carefully.~~

I've combed, refined and made it as good as I can think to for now. 

a semi-technical demonstration of it in action.
https://youtu.be/kl5emKQW0O8

game_initial_spawn.lua and map_startbox.lua are both super super
critical and handshake with many other aspects of the game. Therefore, I
made a _experimental version of each and gate which pair to run behind a
modoption in the experimental tab for testing. The vanilla game after
this PR should be unchanged except for a minor alteration to invalid
start position text (too close to other player)