Finish CATT upgrades (#6289)
Advances PR by Itanthias/KAS: "CATT aiming script now accounts for unit
turning", #5641. Changes are per that PR.

### Work done

Resolved some nitpicks and reverted part of the larger change so that
only `cordoom` and `legeheatraymech` (Bulwark and Sol Invictus) are
excluded from the CATT upgrade to their unit scripts, for now.

These both needed multiple script headers for their multiple,
independently targeting weapons, any of which might want to rotate the
unit's chassis. The better solution imo is to get the simple change into
the base game and tested, then to upgrade Bulwark and Sol together.
That's a desirable change and will happen soon so we won't have these
split files for long enough for them to be a maintenance burden.

---------

Co-authored-by: Kyle Shepherd <kyleanthonyshepherd@gmail.com>
Co-authored-by: efrec <quantumscrewdriver@gmail.com>