(hotfix) Fix aoe widget crash when unit targets itself (#6433)
There was div by 0 when a ship with direct/noexplode weapon targetted
itself.
For example Brimstone, Hippocampus, Thalassa

I changed it so it doesn't try to draw indicator if there is going to be
div by 0.
I also changed it so that indicator does not snap to unit when it target
itself.