fix carriers reentering UnitDestroyed, nil accesses (#6442)
### Work done

Mini patch adding some guards and re-checking that tables are valid
halfway through UnitDestroyed to avoid encountering this bug again.

### Fixes issue

Dead drones issue stockpile orders to their carriers, which triggers the
carrier's UnitCommand, which reaches the UpdateCarrier method. That
method may reenter UnitDestroyed. This is probably some modded garbage
but that's none of my business.

Also, the subUnitID might be cleaned up, and the carrier might be
cleaned up, by the time that we reach each of their UnitDestroyed
methods. When unit deaths happen on the same frame, probably some of the
code's assumptions go out the window. So I added a double-check that the
table is valid. An easier method would be just to keep a copy of the
table local, but that is not how any of this file was written, so I did
not do that.

This is a frequent error in the current logs due to uptick in Legion
games. Silencing it would be nice.

Co-authored-by: efrec <quantumscrewdriver@gmail.com>