Re-Add Resurrection Rework (#6430)
PLAYERS:
The backend of resurrection handling has been reworked. Keep an eye out
for any strange behavior related to resurrection (besides the known
failure to complete rez-repair bug)

NERDS:

This is a revert of the revert - now without the XP gain upon
resurrection. That part will be uncommented in season 3 potentially.

about:

https://youtu.be/yHnD4Jo-tq8

Rewrite the resurrection gadget to transfer data from previous unitID to
new unitID in a GG accessible way. Right away, this allows us to
transfer the XP from the previous unitID to the new unitID and better
handle units that can both build and resurrect - such as scavenger
builders.

EDIT: Post rework, now there's an extra param in the FeatureCreated
callin for "SourceID". If the feature was created from a unit that died,
it will be the unitID of its prior incarnation. corpse_linker.lua awaits
engine replacement. Thanks Sprung for your help.