fix rare issue with tweaks and translatedHumanName (#6468)
Most tweakdefs modders don't encounter this but when they do they get an
error message they don't understand. This is slightly better for widget
compatibility imo.

### Work done

- Guard a nil access when proxyUnitDef does not exist in a custom
Scavenger unit with a bad `fromunit` field. This was just easier to do
than refiguring how we do unit copying -- that's a lot of work.

Co-authored-by: efrec <quantumscrewdriver@gmail.com>