big refactor, minor changes: adds more setup and game state, non-functional commit
There is a lot to change when slotting in new functionality at the moment. Buggeroff is not really the optimal place for some of this code, also, so this commit suffers from a few abuses. The end result should look better and be easier to figure out via the next commit alone, though, which is the point.

The actual changes lean mostly toward marking units as visited differently. Right now, non-allies can mark a unit as visited, which seems wrong/perverse (though almost zero abuse potential in-game). Everything else (I think) is a non-functional change.

Sorry for some of the pointless shuffling. I think it will make more sense in just a moment.