fix: builder buggeroff ignores repeat for immobile build tasks (#6478)
### Work done

Ignore the builder's repeat state when placing a static/structure unit.

### Addresses issue

Very minor issue: It is somewhat common to have Repeat enabled by
accident on builders, since it has very few effects, ordinarily. With
buggeroff, though, it causes the scram behavior to occur only once per
build task, which isn't how the current gadget is designed (it needs
multiple updates to orient units away from the build site).

Using command-style logic in a CommandFallback instead of current
implementation, this type of fix is not necessary, so this change could
be reverted in the future, maybe.

Co-authored-by: efrec <quantumscrewdriver@gmail.com>