fix: area timed damage regressions/merge conflict variety pack (#6484)
These are minor regressions in napalm/acid:

- The damage frame guard on factory-built units was always one frame
long
- Some AI goofs during optimization passes were anti-optimizations
- Some human goofs during a merge conflict regressed the hit detection
on units, so only features are using fancy base position + mid position
math

I wanted to do an actual optimization pass on this and found the file
not as I expected it to be. These are not causing problems in the game
but also are a blocker on further work here.

Tested each of these changes and they are working correctly.

---------

Co-authored-by: efrec <quantumscrewdriver@gmail.com>