add example unit: legadvaabot
The unit has a long-range railgun slaved to a medium-range miniflak. It needs this for its animations/unit script to work correctly, so we cannot just remove the constraint. Instead, we can give the miniflak a massive preaim range by setting the new preaim_range customparam to match its railgun weapon's range.