perf: refactor of area_timed_damage
I net about +6% FPS on the napalm vs heavy units stress test (not enough targets to matter) and about +10% FPS on napalm vs mass spam, and no longer regularly dipped down to x0.90 game speed on either test, which used to be normal. Using the same crappy laptop from before so that seems good.

The overwhelming cost of e.g. napalm goes to its 1000-particle areas of effect so there's very little to do in the Lua except make it much easier to read, which this does.