Fix misspellings in English locales (#6533)
First off - great game! I'm looking to contribute and thought I start
off with some low-hanging fruit while I get my feet wet in the codebase!
<!--
PR Template! Please make sure to give your PR a relevant title so a
squash merge remains descriptive
If any commented sections are not relevant to this PR, remove them.
Please fill out the uncommented sections with any relevant information.
-->

### Work done
<!--
Describe the changes or additions made in this PR, and why they
are necessary or important. If there is unusual complexity in the
code or functionality, please explain it so reviewers can understand.
-->

This is a nonfunctional change to fix some misspellings and grammatical
errors in the English locales. I do not speak any of the other languages
that locales are provided for, so these changes may also need to be
applied there.

<!-- If relevant
#### Addresses Issue(s)
- Issue URL
-->

<!-- If relevant
#### Setup
Describe any setup requirements to test this work (Specific settings,
widgets, etc))
-->

#### Test steps
- [ X] Write the steps needed to test out the changes. Include the
expected result.

These are merely locale string changes - no functional no specific
testing is needed for functionality.

<!-- If relevant
### Screenshots:
If you're making visible changes, add before/after screenshots or videos
of the major
changes so it's easier for reviewers to see what is different in this PR

#### BEFORE:
(screenshot from master)

#### AFTER:
(screenshot from branch)
-->