fix: clamp weapondef proximityPriority to a rational range (#6538)
Basic groundwork for evaluating a change to weapon autotargeting
proximity. Weapons with prox <= -1 were permanently affixing to targets
outside their range and even waffling between them.

### Work done

- Clamps proximity priority to guarantee that targets in range are
always preferable to targets not in range.
- That change is not actually sufficient to achieve that result; a minor
insensitivity to changing the aim angle is also needed. That is a change
to unitDefs post-weaponDefs-post-processing, which isn't really how
things work at the moment. A final follow-up pass would be good though.
So, todo: Add a units-weapons link table and use it during post
processing steps.
- minor code maint

---------

Co-authored-by: efrec <quantumscrewdriver@gmail.com>