Legion Update 12.25.4 (Legion Anti-nuke Vehicle) (#6545)
### Work Done
Adds a temporary unit, the Hera mobile anti-nuke vehicle, to the Legion
T2 vehicle to replace the placeholder Cortex Saviour in the lab. Legion
really needs a non-all-terrain bot anti-nuke which is currently under
development from Mqiib, a junior modeller, so this is a stop gap measure
only and will be axed in the future.