Commando animation/def fixes (#6529)
Players:
The commando's animation script has been improved.
- it will no longer swap rapidly between targets
- no longer shoots sideways
- no longer strains to shoot things just outside its range
- deploys faster, and un-deploys thereafter rather than staying deployed
forever.
- removed an accidentally added 20 energycost per shot for its main
weapon
- it can now be given the Attack command.


nerds:

https://github.com/beyond-all-reason/Beyond-All-Reason/commit/27618a2b5373cfbbeb88c290d60dd0d58f60045d
As an oversight of a laser energycost standardization effort, the
LaserCannon weapon larping as a cannon was given an unfitting energycost
by accident. removed it.
<img width="1489" height="869" alt="image"
src="https://github.com/user-attachments/assets/5c09f465-6df8-423b-91df-f60826cbf01a"
/>
Website lists it as a cannon and only LRPC's cost energy to fire in that
category

aim scripts are black magic these days. I figured out a way to allow
shots based on timestamping periodically the heading and if theyre close
enough to the same allow it to shoot.
Tested and seems to be working fine